requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
psycopg[binary]==3.2.10
qdrant-client==1.14.2
redis==6.4.0
pydantic-settings==2.9.1
httpx==0.28.1
pytest==8.4.1
ruff==0.12.8
